Extracted and renamed `ReplicationsCompletedListener`
Moved `ReplicationPolicyManager.ReplicationsCompletedListener` ->
`PolicyReplicationsCompletedListener`.
Moved `ReplicationPolicyManager.SimpleReplicationsCompleteListener` ->
`PolicyReplicationsCompletedListener.SimpleListener`.
