+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/replication/ReplicationPolicyManager.java
@@ -29,27 +29,6 @@ public class ReplicationPolicyManager {
     private final List<Replicator> replicators = new ArrayList<Replicator>();
     private final ReplicationListener replicationListener = new ReplicationListener();
 
-    /**
-     * Interface with methods to receive callbacks for different termination states
-     * of the replications configured by a replication policy.
-     */
-    public interface ReplicationsCompletedListener {
-        /**
-         * Gets called back when all replicators completed.
-         */
-        void allReplicationsCompleted();
-        /**
-         * Gets called back when a replication completed.
-         * @param id the replication ID
-         */
-        void replicationCompleted(int id);
-        /**
-         * Gets called back when a replication has an error.
-         * @param id the replication ID
-         */
-        void replicationErrored(int id);
-    }
-
     protected void startReplications() {
         synchronized (replicators) {
             for (Replicator replicator : replicators) {
@@ -78,32 +57,13 @@ public class ReplicationPolicyManager {
 
     /**
      * Interested parties wishing to receive replication lifecycle events should call
-     * this method with a class implementing the {@link ReplicationsCompletedListener} interface.
+     * this method with a class implementing the {@link PolicyReplicationsCompletedListener} interface.
      *
      * @param listener A class implementing appropriate callback methods for replication lifecycle
      *                 events
      */
-    public void setReplicationsCompletedListener(ReplicationsCompletedListener listener) {
+    public void setReplicationsCompletedListener(PolicyReplicationsCompletedListener listener) {
         replicationListener.setReplicationsCompletedListener(listener);
     }
 
-    /**
-     * A simple {@link ReplicationsCompletedListener}
-     * to save clients having to override every method if they are only interested in a subset of
-     * the events.
-     */
-    public static class SimpleReplicationsCompletedListener implements
-            ReplicationsCompletedListener {
-        @Override
-        public void allReplicationsCompleted() {
-        }
-
-        @Override
-        public void replicationCompleted(int id) {
-        }
-
-        @Override
-        public void replicationErrored(int id) {
-        }
-    }
 }
